Make toggling 'Show template' on/off faster
