+++ packages/edit-site/src/components/block-editor/index.js
@@ -10,12 +10,11 @@ import { privateApis as editPatternsPrivateApis } from '@wordpress/patterns';
 import TemplatePartConverter from '../template-part-converter';
 import { SidebarInspectorFill } from '../sidebar-edit-mode';
 import SiteEditorCanvas from './site-editor-canvas';
-import useBlockEditorProvider from './use-block-editor-provider';
+import BlockEditorProvider from './block-editor-provider';
 
 import { unlock } from '../../lock-unlock';
 const { PatternsMenuItems } = unlock( editPatternsPrivateApis );
 export default function BlockEditor() {
-	const BlockEditorProvider = useBlockEditorProvider();
 	return (
 		<BlockEditorProvider>
 			<TemplatePartConverter />
