+++ packages/edit-site/src/components/block-editor/providers/default-block-editor-provider.js
@@ -4,6 +4,8 @@
 import { useEntityBlockEditor } from '@wordpress/core-data';
 import { privateApis as blockEditorPrivateApis } from '@wordpress/block-editor';
 import { useSelect } from '@wordpress/data';
+import { useMemo } from '@wordpress/element';
+import { createBlock } from '@wordpress/blocks';
 
 /**
  * Internal dependencies
@@ -14,7 +16,12 @@ import useSiteEditorSettings from '../use-site-editor-settings';
 
 const { ExperimentalBlockEditorProvider } = unlock( blockEditorPrivateApis );
 
-export default function DefaultBlockEditorProvider( { children } ) {
+const noop = () => {};
+
+export default function DefaultBlockEditorProvider( {
+	contentOnly,
+	children,
+} ) {
 	const settings = useSiteEditorSettings();
 
 	const { templateType } = useSelect( ( select ) => {
@@ -25,17 +32,39 @@ export default function DefaultBlockEditorProvider( { children } ) {
 		};
 	}, [] );
 
-	const [ blocks, onInput, onChange ] = useEntityBlockEditor(
+	const [ entityBlocks, onInput, onChange ] = useEntityBlockEditor(
 		'postType',
 		templateType
 	);
 
+	const contentOnlyBlocks = useMemo( () => {
+		return [
+			createBlock(
+				'core/group',
+				{
+					layout: { type: 'constrained' },
+					style: {
+						spacing: {
+							margin: {
+								top: '4em',
+							},
+						},
+					},
+				},
+				[
+					createBlock( 'core/post-title' ),
+					createBlock( 'core/post-content' ),
+				]
+			),
+		];
+	}, [] );
+
 	return (
 		<ExperimentalBlockEditorProvider
 			settings={ settings }
-			value={ blocks }
-			onInput={ onInput }
-			onChange={ onChange }
+			value={ contentOnly ? contentOnlyBlocks : entityBlocks }
+			onInput={ contentOnly ? noop : onInput }
+			onChange={ contentOnly ? noop : onChange }
 			useSubRegistry={ false }
 		>
 			{ children }
